add rails 5.2
