+++ .github/workflows/ci.yml
@@ -13,11 +13,15 @@ jobs:
       matrix:
         ruby: [2.7, '3.0']
         gemfile:
-          - rails_6.0
-          - rails_6.1
           - rails_7.0
+          - rails_6.1
+          - rails_6.0
+          - rails_5.2
+        exclude:
+          - ruby: '3.0'
+            gemfile: rails_5.2
     runs-on: ubuntu-latest
-    env: # $BUNDLE_GEMFILE must be set at the job level, so it is set for all steps
+    env:
       BUNDLE_GEMFILE: ${{ github.workspace }}/gemfiles/${{ matrix.gemfile }}.gemfile
     steps:
     - uses: actions/checkout@v2
@@ -25,4 +29,4 @@ jobs:
       with:
         ruby-version: ${{ matrix.ruby }}
         bundler-cache: true # runs 'bundle install' and caches installed gems automatically
-    - run: bundle exec rake
+    - run: bundle exec rake db:create db:schema:load test
