+++ gemfiles/rails_5.2.gemfile
@@ -1,0 +1,17 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "cells-erb"
+gem "cells-rails"
+gem "reform-rails"
+gem "trailblazer"
+gem "trailblazer-cells"
+gem "trailblazer-loader"
+gem "sqlite3", "~> 1.3.8"
+gem "minitest-capybara"
+gem "appraisal"
+gem "actionpack", "~> 5.2.0"
+gem "activerecord", "~> 5.2.0"
+
+gemspec path: "../"
