+++ gemfiles/rails_5.2.gemfile.lock
@@ -8,35 +8,37 @@ PATH
 GEM
   remote: https://rubygems.org/
   specs:
-    actionpack (7.0.0.rc1)
-      actionview (= 7.0.0.rc1)
-      activesupport (= 7.0.0.rc1)
-      rack (~> 2.0, >= 2.2.0)
+    actionpack (5.2.6)
+      actionview (= 5.2.6)
+      activesupport (= 5.2.6)
+      rack (~> 2.0, >= 2.0.8)
       rack-test (>= 0.6.3)
       rails-dom-testing (~> 2.0)
-      rails-html-sanitizer (~> 1.0, >= 1.2.0)
-    actionview (7.0.0.rc1)
-      activesupport (= 7.0.0.rc1)
+      rails-html-sanitizer (~> 1.0, >= 1.0.2)
+    actionview (5.2.6)
+      activesupport (= 5.2.6)
       builder (~> 3.1)
       erubi (~> 1.4)
       rails-dom-testing (~> 2.0)
-      rails-html-sanitizer (~> 1.1, >= 1.2.0)
-    activemodel (7.0.0.rc1)
-      activesupport (= 7.0.0.rc1)
-    activerecord (7.0.0.rc1)
-      activemodel (= 7.0.0.rc1)
-      activesupport (= 7.0.0.rc1)
-    activesupport (7.0.0.rc1)
+      rails-html-sanitizer (~> 1.0, >= 1.0.3)
+    activemodel (5.2.6)
+      activesupport (= 5.2.6)
+    activerecord (5.2.6)
+      activemodel (= 5.2.6)
+      activesupport (= 5.2.6)
+      arel (>= 9.0)
+    activesupport (5.2.6)
       concurrent-ruby (~> 1.0, >= 1.0.2)
-      i18n (>= 1.6, < 2)
-      minitest (>= 5.1)
-      tzinfo (~> 2.0)
+      i18n (>= 0.7, < 2)
+      minitest (~> 5.1)
+      tzinfo (~> 1.1)
     addressable (2.8.0)
       public_suffix (>= 2.0.2, < 5.0)
     appraisal (2.4.1)
       bundler
       rake
       thor (>= 0.14.0)
+    arel (9.0.0)
     builder (3.2.4)
     capybara (3.36.0)
       addressable
@@ -97,13 +99,12 @@ GEM
       nokogiri (>= 1.6)
     rails-html-sanitizer (1.4.2)
       loofah (~> 2.3)
-    railties (7.0.0.rc1)
-      actionpack (= 7.0.0.rc1)
-      activesupport (= 7.0.0.rc1)
+    railties (5.2.6)
+      actionpack (= 5.2.6)
+      activesupport (= 5.2.6)
       method_source
-      rake (>= 12.2)
-      thor (~> 1.0)
-      zeitwerk (~> 2.5)
+      rake (>= 0.8.7)
+      thor (>= 0.19.0, < 2.0)
     rake (13.0.6)
     reform (2.6.1)
       disposable (>= 0.5.0, < 1.0.0)
@@ -117,9 +118,10 @@ GEM
       declarative (< 0.1.0)
       trailblazer-option (>= 0.1.1, < 0.2.0)
       uber (< 0.2.0)
-    sqlite3 (1.4.2)
+    sqlite3 (1.3.13)
     temple (0.8.2)
     thor (1.1.0)
+    thread_safe (0.3.6)
     tilt (2.0.10)
     trailblazer (2.1.0)
       trailblazer-macro (>= 2.1.0, < 2.2.0)
@@ -151,19 +153,18 @@ GEM
       trailblazer-activity-dsl-linear (>= 0.4.1, < 1.0.0)
       trailblazer-developer (>= 0.0.21, < 1.0.0)
     trailblazer-option (0.1.2)
-    tzinfo (2.0.4)
-      concurrent-ruby (~> 1.0)
+    tzinfo (1.2.9)
+      thread_safe (~> 0.1)
     uber (0.1.0)
     xpath (3.2.0)
       nokogiri (~> 1.8)
-    zeitwerk (2.5.1)
 
 PLATFORMS
   x86_64-linux
 
 DEPENDENCIES
-  actionpack (~> 7.0.0.rc1)
-  activerecord (~> 7.0.0.rc1)
+  actionpack (~> 5.2.0)
+  activerecord (~> 5.2.0)
   appraisal
   bundler
   cells-erb
@@ -172,7 +173,7 @@ DEPENDENCIES
   minitest-capybara
   rake
   reform-rails
-  sqlite3
+  sqlite3 (~> 1.3.8)
   trailblazer
   trailblazer-cells
   trailblazer-loader
