--- lib/trailblazer/rails/test/integration.rb
@@ -1,6 +1,0 @@
-require "minitest/rails/capybara" # loads Capybara, etc.
-
-module Trailblazer::Test
-  class Integration < Capybara::Rails::TestCase
-  end
-end
