--- test/dummy/Gemfile
@@ -1,17 +1,0 @@
-source "https://rubygems.org"
-
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem "actionpack", "~> 6.1.0"
-gem "activerecord", "~> 6.1.0"
-# Use sqlite3 as the database for Active Record
-gem "sqlite3"
-
-eval_gemfile("../../Gemfile")
-
-group :test do
-  gem "minitest-capybara"
-end
-
-gem "minitest-line"
-
-gem "byebug"
