--- test/dummy/Rakefile
@@ -1,6 +1,0 @@
-# Add your own tasks in files placed in lib/tasks ending in .rake,
-# for example lib/tasks/capistrano.rake, and they will automatically be available to Rake.
-
-require File.expand_path("config/application", __dir__)
-
-Rails.application.load_tasks
