+++ test/dummy/app/concepts/song/operation/create.rb
@@ -1,8 +1,3 @@
-class Song::Operation::New < Trailblazer::Operation
-  step Model(Song, :new)
-  step Contract::Build(constant: Song::Contract::Create)
-end
-
 class Song::Operation::Create < Trailblazer::Operation
   step Subprocess(Song::Operation::New)
   step Contract::Validate()
