+++ test/dummy/app/concepts/song/operation/new.rb
@@ -1,0 +1,4 @@
+class Song::Operation::New < Trailblazer::Operation
+  step Model(Song, :new)
+  step Contract::Build(constant: Song::Contract::Create)
+end
