--- test/dummy/bin/bundle
@@ -1,4 +1,0 @@
-#!/usr/bin/env ruby
-
-ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
-load Gem.bin_path("bundler", "bundle")
