--- test/dummy/bin/rails
@@ -1,5 +1,0 @@
-#!/usr/bin/env ruby
-
-APP_PATH = File.expand_path("../config/application", __dir__)
-require_relative "../config/boot"
-require "rails/commands"
