--- test/dummy/bin/rake
@@ -1,5 +1,0 @@
-#!/usr/bin/env ruby
-
-require_relative "../config/boot"
-require "rake"
-Rake.application.run
