+++ test/dummy/config/boot.rb
@@ -1,3 +1,1 @@
-ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
-
 require "bundler/setup" # Set up gems listed in the Gemfile.
