--- test/dummy/config/initializers/filter_parameter_logging.rb
@@ -1,4 +1,0 @@
-# Be sure to restart your server when you modify this file.
-
-# Configure sensitive parameters which will be filtered from the log file.
-Rails.application.config.filter_parameters += [:password]
