--- test/dummy/config/initializers/mime_types.rb
@@ -1,4 +1,0 @@
-# Be sure to restart your server when you modify this file.
-
-# Add new mime types for use in respond_to blocks:
-# Mime::Type.register "text/richtext", :rtf
