--- test/dummy/config/initializers/session_store.rb
@@ -1,3 +1,0 @@
-# Be sure to restart your server when you modify this file.
-
-Rails.application.config.session_store :cookie_store, key: "_rails_session"
