--- test/dummy/config/initializers/sqlite3_adapter.rb
@@ -1,1 +1,0 @@
-# Rails.application.config.active_record.sqlite3.represent_boolean_as_integer = true
