+++ test/dummy/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20_161_122_145_124) do
+ActiveRecord::Schema.define(version: 1) do
   create_table "songs", force: :cascade do |t|
     t.string "title"
   end
